+++ src/peds/Ped.cpp
@@ -2854,7 +2854,7 @@ CPed::SetObjective(eObjective newObj, void *entity)
 		case OBJECTIVE_WAIT_IN_CAR_THEN_GETOUT:
 
 			// In this special case, entity parameter isn't CEntity, but int.
-			SetObjectiveTimer((int)entity);
+			SetObjectiveTimer((uintptr)entity);
 			break;
 		case OBJECTIVE_KILL_CHAR_ON_FOOT:
 		case OBJECTIVE_KILL_CHAR_ANY_MEANS:
@@ -3347,7 +3347,7 @@ CPed::MakePhonecall(void)
 		if (!IsPlayer() && CTimer::GetTimeInMilliseconds() > m_phoneTalkTimer - 7000 && bRunningToPhone) {
 
 			FindPlayerPed()->m_pWanted->RegisterCrime_Immediately(m_crimeToReportOnPhone, GetPosition(),
-				(m_crimeToReportOnPhone == CRIME_POSSESSION_GUN ? (int)m_threatEntity : (int)((CPed*)m_pEventEntity)->m_threatEntity), false);
+				(m_crimeToReportOnPhone == CRIME_POSSESSION_GUN ? (uintptr)m_threatEntity : (uintptr)((CPed*)m_pEventEntity)->m_threatEntity), false);
 			bRunningToPhone = false;
 		}
 #endif
@@ -4555,7 +4555,7 @@ CPed::SetPointGunAt(CEntity *to)
 		SetLookFlag(to, true);
 		SetAimFlag(to);
 #ifdef VC_PED_PORTS
-		SetLookTimer(INT_MAX);
+		SetLookTimer(INT32_MAX);
 #endif
 	}
 
@@ -4794,8 +4794,8 @@ CPed::SetEvasiveDive(CPhysical *reason, uint8 onlyRandomJump)
 	if (reason->IsVehicle() && m_nPedType == PEDTYPE_COP) {
 		if (veh->pDriver && veh->pDriver->IsPlayer()) {
 			CWanted *wanted = FindPlayerPed()->m_pWanted;
-			wanted->RegisterCrime_Immediately(CRIME_RECKLESS_DRIVING, GetPosition(), (int)this, false);
-			wanted->RegisterCrime_Immediately(CRIME_SPEEDING, GetPosition(), (int)this, false);
+			wanted->RegisterCrime_Immediately(CRIME_RECKLESS_DRIVING, GetPosition(), (uintptr)this, false);
+			wanted->RegisterCrime_Immediately(CRIME_SPEEDING, GetPosition(), (uintptr)this, false);
 		}
 	}
 }
